requirements-ndi.txt
requests
websockets
Flask
flaskwebgui>1.0
psycopg2-binary
mysql.connector
pillow
waitress

# Does not work with Python >3.10
# Required for NDI support
opencv-python
ndi-python
numpy==1.25.2